Pick default activity based on device properties
A single APK could have multiple launcher activities. So far, we've
only checked if one of those has the DEFAULT_CATEGORY, and if so,
given preference to that.

However, if the apk has a leanback launcher, and a regular launcher
(for a phone, say), we should pick the default based on the target.
This CL looks at the target device's hardware characteristics, and
if its a TV, then prefers the leanback launcher.

This CL also moves all the activity location logic out of
AndroidUtils, and moves them mostly into DefaultActivityLocator.

Fixes https://code.google.com/p/android/issues/detail?id=176033

Change-Id: Iccfd5f74c575272bc8a45fb72fe85d780ccfa3ee