Fix "@id/" resource parsing in layout files.
Parse all attributes to search for ids declared on the fly. This is
important if the @+id/ declaration is not in 'android:id' attribute. Try
to mark the source of the id as the tag using the id in 'android:id'
attribute. If such a tag is not found, add the id with the source as the
first tag referencing the id.

Also fix style warnings in the file.

Bug: http://b.android.com/172239
Change-Id: I77295040bed76e79a785e71fbbbb3a376bb9ba98