Translations editor UI test: fix incorrect condition
In a previous attempt to fix this flaky test, we assumed
that checking for whether the loading panel is complete
would be sufficient. However, that fails if the check returns
even before the loading panel has begun loading. So now
we check that the table has been initialized with the
proper model.

Change-Id: Ic0aaa3e8521016ea5cdb10e8c1eb1f00345a2f4d