[UITests] First step towards reporting ignored/skipped tests.
Currently GuiTestRunner gets to decide on-the-fly whether to skip certain
tests:
  1. if there is no DISPLAY attached (headless environment)
  2. if an IDE internal error occurred (likely due to a previous test)
  3. if the test requires a different JDK version.
(additionally, JUnit will honor the @Ignored annotation.)

The problem is that we don't notify the framework, and all the above appear
as successful executions (in <0.1s). ~80 tests in our latest Jenkins run
did not in fact execute.

This CL allows us to distinguish #1 and #2 above from passing tests.

Change-Id: I3716d77163b2f406b3efa7ee24b4980585ea06c1