Adding GPU monitor to Android Studio.
This monitor reports the breakdown of time used to render a frame.

Pre-M, it is broken down into four timings groups that sum up to the
final frame render time:
{Draw, Prepare, Process, Execute}.
Also, since pre-M stats don't provide timestamps, the values are
spread evenly over the sampling period. The sample frequency is turned
up to account for this.

In M and above, the timings are broken down into nine groups:
{vsync delay, input handling, animations, layout, draw, sync,
 command issuance, swap buffers, miscellaneous}

Change-Id: I40db8a47a2072138da13d5439e04faf9de7bf693