Nele: Change Drag and Drop format to contain multiple components.
This is in preparation for drag and drop in the structure pane.
Change the Drag and Drop format currently only used by the palette
to be able to contain multiple components. Also add information to
determine the type of drop and if the source model is identical to
the target model.

Change-Id: I219e3a316c180613fc1b7cc24d0144159867e5f2