Nele: Change the approach for updating the structure tree after a model change.
The previous code was keeping references to XmlTag references which is a memory
leak in itself because they were never discarded. Worse: the instances were
being reused such that the tree would expand the wrong nodes after an update.
This has now been fixed by holding a map from component id to tree node and
attempting to restore the visibility of all the nodes we recognize the id for.
[This may cause some nodes without id to close and other nodes with duplicate
ids to open - but these issues are deemed less important that the previous
problems.]

Change-Id: I00818473add6dcac10c658b1d191110ff8584c40