Moved wizard UI components into ui package
Various UI components lived in the wizard package for, what seems
like, convenience, as they were only used by initial wizard code.
However, there is a ui package which should really be the home for
classes like that, and as a bonus, it cleans up the wizard package,
which became cluttered with all sorts of extra classes over time.

Change-Id: I28ddd5f9db593e63f0a60d3f830023361bd88e3c