Fixing warnings during studio exit caused by sdk updates
When installing updates to sdk packages on windows there is an
option to exit studio and use the old sdk manager to avoid
trying to replace in-use files. When doing so there was an
assertion error since we tried to run the installation wizard
without setting it up properly first (since we were already
exiting). This change causes the wizard not to try to run in
this case.
There were also a couple warnings about components not being
disposed properly on exit. One was the wizard host in this case,
and one was because we weren't starting and stopping the progress
window correctly during synchronous sdk refresh. Fixing those as
well.

Change-Id: Ic6f9900de31462afa2f7e5ab471e238884fb418d