[UITests] Canonicalize project creation path.
Previously this broke tests on MacOS because the default tmpdir
is a symlink. Also, test execution resulted in a flurry of empty
folders created in there, because getProjectCreationDirPath had the
side effect of creating a fresh tmp dir every time.

This change creates a single tmp path and canonicalizes it right away.
getProjectCreationDirPath() is now side-effect free.

Further investigation is necessary in order to remove other calls to
toCanonicalPath() which is symlink-unfriendly.

Change-Id: Ic6f3c456f4320847d1a3ec44a0878c976cbf52cc