Renamed AndroidFacet#isGradleProject to "requiresAndroidModel".
This change communicates better that a project is backed by an
AndroidProject obtained from the build system. Finding a good
name was not easy. I originally renamed the method to
"hasAndroidModel", but it is easy to misunderstand and assume
that it means getAndroidModel() != null, when in fact, the
project can require an Android model but the model be null
(e.g. if sync fails.)

Additional changes:
- Renamed AndroidFacet#getIdeaAndroidProject to "getAndroidModel"
- Renamed AndroidFacet#setIdeaAndroidProject to "setAndroidModel"
- Renamed Projects#isGradleProject to "requiresAndroidModel"
- Replaced "Gradle project" with "project backed by an
  AndroidProject" in Javadocs
- Renamed IdeaAndroidProject#getDelegate to
  "getAndroidProject"

Bug: 22597272

Change-Id: I6e9e5ff25e3844b75f5635898a151ef33682f663