Fixing creation of wear AVDs on windows
In a device definition, a skin can be specified. Then when we create an
AVD we look for a skin with that name in a few places, including in
system images.
Previously, in the system image check, we were using '/' as a path
seperator explicitly, which caused a problem on Windows for such devices
(in practice, wear devices).

Change-Id: I408119e8cc752d7fc487d373cce2abc42bf9d052