Only setting height to 1 when we are reducing the width
This was part of the original patch but, since layoutlib was fixed, it
wasn't necessary anymore. However, I ran into the issue today again when
using an old layoutlib so I think it is safer to add it right now.
When passing 1 to the height, in some cases, layoutlib will not
correctly recalculate the size and will be stuck at a height 0, causing
it to stop rendering.

Change-Id: If482d688d1a29e2f948262049bcfd4cad9cb36fe