Add variant selector to the theme parent.
The variant selector will display the different values for the different
qualifiers only if the theme is defined in multiple folders. This also
makes the parent cell UI look more like the Color/Drawable editor.
This CL creates a ConfiguredElement class that allows associating a
value (an attribute or a theme parent) to a FolderConfiguration so it
can be passed around with the FolderConfiguration attached.

The CL also removes a few warnings and adds the separator for the table.

Change-Id: Iaaa5e0b86e86e074f56ce4c11870c0ec06084c1e