Fix IllegalArgumentException log in resolveDrawable
ResourceHelper#createStateListState spams log with errors when
legitimate state lists are parsed. This CL fixes that, clarifies
documentation of ResourceHelper#createStateListItem and makes
it return null when it can't parse an item.

Easiest way to reproduce error: go to AndroidManifest.xml of default
project and press Ctrl+Q when cursor is on "@style/AppTheme"
reference.

For an example of state list that isn't parsed, look
fastscroll_thumb_material.xml in the Android framework.

Change-Id: I5eedafb89a4b772673af59ff4559a230dfdba829