Allow relative paths in template recipe files.
Setup for template reuse by allowing a recipe file to reference
files outside of its own subdirectory.
This is accomplished by setting the template folder from
FreemarkerUtils and restoring it upon completion of the processing.
If the caller may load additional files based on the loaded xml,
then the caller must do this work in a callback.

Also: disable localized lookup in FreeMarker.
The template system used to look for Java style language overrides.

Change-Id: Ie5121817a6473a55133981fe27e7e1c009e0d8d0