Implement the Gradle Java Dsl Parser.
Generalize the parser API so it would take the DSL element that
the parser would modify.
Add a map to hold the extra data provided by extra parser
(e.g. Java DSL parser), and a API to get that data.
Note: the mechnism to load extra parser from extension havn't been
implemented yet and it will be in a seperate CL.

Add the Java DSL parser that could currently handle top level
sourceCompability and targetCompability assignment.

Change-Id: I8a366f65726add952041e9683aaf1b70ed5fe0a0