Do not re-examine jars when loading classes.
Fix a performance issue where all of a module's transitive external jars
were re-examined for each module in its transitive module dependencies.
Instead, look at all the transitive jars once. This reduces preview
loading time from > 20 minutes to < 30 seconds for one client project.

Note that as written, this change DOES alter the order that modules/jars
are searched for the target class.
Previously, the order was:
1. The module itself.
2. All the transitive external jars.
3. All the transitive dependent modules.

Now, it is:
1. The module itself and all its transitive dependent modules.
3. All the transitive external jars.

Change-Id: Ic1d839760e74c77bf82f45f598dab6704c86ddf7