Fixes ThemeResolver and parent resolution
The ThemeResolver wasn't being updated in configuration changes which
produced inconsistent results when closing and re-opening the theme
editor. Now, the ThemeResolver listens for configuration changes and
refreshes itself.
Also stop resolving themes in getAllParents. When using a configuration,
the name resolution was failing when a parent didn't exist in certain
API level. For example, when using API 19, the getAllParents would try
to get all the available parents, including those above API 19, failing
to resolve those.

Change-Id: If98f6bd50cd1fc32d78d2bedae154ada6cd208bd