Fixes to editing SDK path in SDK manager
Previously, if you changed the SDK path in the SDK manager to
an existing location that contained a different (either newer
or older) version of a component that the sdk setup step wants,
it would try to download the version it wanted and fail.
In fact when you're in the sdk manager and changing the sdk path,
we don't want the process of changing it to download new components,
unless you explicitly request it using the checkboxes.

There was also a problem where every time it would show a popup
saying that you didn't have an sdk. Moved that code to the welcome
screen provider so it only shows at the correct time.

Change-Id: Id58c0722f0095efb4ffbbc5e640d0a3f051bf309