Skip equal line locations when committing
A GridLayout with two children at (0, 0) and (0, 9) looks like it only has
three columns. When I dragged the right child to the middle column the drag
handler dutifully set its location to (0, 8) but the preview didn't change.
That was unintuitive.

It turned out that the GridLayout in its initial state had many equal line
locations that messed up the calculation of the final column index.

Change-Id: I6a92d0dac51abf30685f06b52c5fcafaadefc6a8