Show details for render errors.
All unexpected errors only showed the exception message and didn't give
the user any way to get the stacktrace. Change the default message to
include a link to show the full stacktrace. This still preserves all the
other formatting that is done - filtering of stacktraces, showing them
inline or removing them entirely etc.

The most common use case for this is when there's an exception in custom
view's onDraw, onMeasure or onLayout method. The old code only showed
the message "onDraw error" with no recourse. After this change, the
message is "onDraw error (Details)" which can be clicked to show the
complete stacktrace.

Change-Id: I4e7dc5caaa214e1ab37cd7dbd1d100ce4d59a292