Revert "Revert "Show details for render errors.""
This reverts commit 7cb68fd90d2d4a8c8cd1a041ccd27951807365b0.

In addition to the revert, this change fixes the tests and makes sure
that the error messages are expanded when they should have been. A new
method is added to the RenderProblem to indicate if the html text in the
message is the default (same as throwable's message).

Also re-enable the testWrongClassFormat after fixing it.

Change-Id: I948b76775371da559b194614b28cfc170680c80e