Revert "Revert "Show details for render errors.""
This reverts commit 48ef3276dd4351a01dd2550f97e7048fc55555e5.

In addition to the revert, this change fixes the tests and makes sure
that the error messages are expanded when they should have been. A new
method is added to the RenderProblem to indicate if the html text in the
message is the default (same as throwable's message).

Also re-enable the testWrongClassFormat after fixing it.

Change-Id: I948b76775371da559b194614b28cfc170680c80e