Fixing NewProjectWizardTest
In the real world we never call createProject with a null project.
Changing the test to use the project from the test fixture.
Not reverting previous temporary fix, since it's required for
TraverseUIStarter. Ideally TraverseUIStarter would run with
an actual project, since it seems like there will be many
inaccuracies as it currently is, but that would be a larger fix.

Change-Id: Ic0bebbb14b180bdfc914b6f2eb1f311796c067cc