Show permission changes on developer services UI
A developer service MAY include an AndroidManifest.xml, which MAY
introduce permissions, e.g.

AndroidManifest.xml
   ...
   <uses-permission android:name="android.permission.INTERNET" />
   <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
   ...

We now parse any AndroidManifest.xml we come across and surface any
permissions we find.

(There was actually a UI hook providing support for this since the
feature was first released but only now are we parsing manifest files
and using it.)

Change-Id: I9fd070fe387a1dc5b0a54dee6e43b21b76d3c754