Don't add empty ids to views.
When in-place editing is used (double clicking a view in the editor),
and no id is provided, it inserted an empty id ('android:id="@+id/"')
in the XML. Fix the issue by only adding the id when the supplied id is
of non-zero length. This also ensures that when removing the id, the
attribute is completely removed.

Bug: http://b.android.com/177811
Change-Id: I512ed12f19632d31f80b5d0efb10c53332affc5a