String in lock file is null-terminated, need to trim
In Java 7 ProcessBuilder starts checking for null characters in
process arguments. The lock file written by the emulator consists
of the emulator PID and, for some reason, a null character, which
was causing ProcessBuilder to barf, and we were silently assuming
the emulator was running.
Trimming the pid.
Also added logging for this situation.
Fixes https://code.google.com/p/android/issues/detail?id=182851

Change-Id: Iee44a7617f849c74067640bdf816687e41550166