Misc code cleanup in the fast deploy manager, and display stats
This CL updates the FastDeployManager to handle resource changes much
more efficiently (instead of a deep codepath which performs file
merging, aapt execution and byte result extraction for each changed
file, it computes these in separate phases such that there's a single
aapt run in the middle which is much more efficient.

It also displays the instant run time: It starts counting milliseconds
and stops when it has received a message back from the app process
that it has finished applying resources and verifying dex changes. (It
doesn't yet count the time to do warm start (typically very fast) or a
cold start (typically not.)

The stats are shown in a balloon; you can disable this with the system
property fd.stats.

Finally, lots of code cleanup (extracting the various path and task
lookups into separate methods which should eventually be fetching
accurate data from the model.

Change-Id: I4f91e71b51f17f00dedc333cc9998630473ab1ce