Use custom running states for cloud configs.
Eliminate the CloudTargetChooser, and instead create running states
specifically for cloud device launch and cloud matrix tests.
Almost all of the state and logic in AndroidRunningState is irrelevant
to these cloud launches, and it is simpler to keep the code separate.

This sets up for moving device selection out of AndroidRunningState
execution.

Change-Id: I6926d3e8ca6f8c7d831d4260b7c610c752bec670