Prevents re-attaching to an already-attached debug session.
This CL prevents the debugger from re-attaching to an already-attached
debug session. This is done by performing an extended scan across all
open projects as well as all live debug sessions. When a match is
found, the host project window will be brought to the foreground and
the debug process's tool window will be activated. This also fixes the
related error when the user attempted to re-attach.

Bug: 75757
Change-Id: I13f66f7457fb4c3918f9405e0aa2876b8640213e