Cleaning up the logcat pipeline
This removes a bunch of buffering, some parsing and a thread, and
instead connects the multiline android receiver straight to the message
input of the console.

This cuts the time to load the initial log from 14s down to less than
.5 seconds on my phone and laptop, and also improves the latency of new
messages arriving to something too small for me to measure.

Fixes http://b.android.com/82775
    Logcat sometimes takes a really long time before logs appear

Change-Id: Ie7d0facb3246d52cdc78d4bfe2deb610ac9b8f33