160901: Provide more control over application installation
Currently when Studio detects that an APK hasn't changed, it
skips installing that APK. Instead, it simply kills the running
app, and re-launches the app.

Although this works in the majority of cases, some users do not
want the app to be killed and just restarted. This CL makes a
small change to the run config UI:
 - logcat settings were simplified: there are only 2 of them now
   Instead of a logcat tab, this is now a logcat group within
   a Misc tab.
 - In the Misc tab, we now have an additional group with an
   "Installation options" group. This exposes 2 check boxes via
   which users can control whether the APK installation can be
   skipped, and whether the app can be killed.

Change-Id: I14825914b0621a0d368f93b694d074918c533297