Prevent the Run window from popping up every time an app is ran.
Currently, AndroidDebugRunner always activates the tool window. If it
was already being shown, then it shows a notification instead. The
first part of this CL changes this behavior: we never show the
toolwindow, and always show the notification.

Since we never show the toolwindow, errors could go unseen. So we
change the process handler to show the tool window only if there are
errors.

Change-Id: Ic32c9d8e94069b9467e8f11720a9b32c05cfeff2