Remove "appCompat" & "appCombatActivity" as template parameters.
The current code uses the wrong sdk version to compute the value
for "appCompatActivity" when creating a new project.
In addition these parameters have to be supplied several places
in the code: old and new template wizards & new project versus
new activity.

Solution: Move the logic into the template.
Add a FM (FreeMarker) method, that can be called while
processing a template, to  determine if the theme used in an
existing module is derived from an AppCompat theme.

This will allow the templates to be used in both the new
and the old activity wizards.

Change-Id: I4304644e1356d3d770a0686f10a00fdd20ad4e89