Temporary fix for updating preview when modifying state list
Since layoutlib reads the value of color state lists from disk,
one needs to make sure the changes to such files are saved as
soon as they are made.

This is not a permanent solution: layoutlib needs to be updated
to be able to handle changes in the PSI files of color state lists.

Bug: http://b.android.com/183767
Change-Id: I22455d851af94ecc43e0bdff671cfc7bac53b70a