Fix the gallery step to include the correct checks.
The minSDK checks did not work during "new project".
Made sure "add activity" and "new project" sets the
same key that the gallery is checking.
Also add checks the existince of an application theme
if required by the selected activity and this is not
for a "new project".

Change-Id: I0fb1f847a8bb9116a6518af34e81936ab02617ab
(cherry picked from commit 27ba27241b0162ab02982600871ff7f2b77cb216)