149270: Fix startup crashing related to missing crypto extensions
For unknown reasons, a number of users have JRE installations
where creating an SSL connection aborts with an error from
javax.crypto. (One possible cause of this scenario is in-place
upgrades of OSX; reinstalling the JDK fixes the problem.)

Unfortunately, for new users, we attempt to make an SSL connection
as part of the Welcome Wizard startup, so when this error happens,
it takes down the whole IDE and users can't start Studio!

We can gracefully recover from this, so let's do that; logging
the exception but letting the user resume.

Change-Id: I504ba89eb34bedd28e41c5c1c436cc2cee6778c9