Provide API for tracking a set of name-values
In certain cases, it is not enough to just track an
event with a label and value, and it is necessary to send
a list of parameter values. GA doesn't provide very good
support for such a scheme, especially if you need to do
queries based on those values.

This CL adds a generic API which allows you to send an
arbitrary list of key-values. However, we don't expose this
generic API as is, and we expose just the specific use case
as necessary for Gradle right now, just so that this API
isn't used all over the place.

Change-Id: I1a592d28dc163f3b7d49f18821cdc8e47e84c1df
(cherry picked from commit a875c22c36aa0dca013ef683745105c22d9e6726)