Change the green color in wizard headers so it does not collide with new art.
We have new icons for activities.
Their green collides with the green on the top of a wizard.
Change the color to a gray to avoid this clash.
Also include the images for "New Module" here.

Change-Id: I1bfbf952dd39ee88a2cfa97e8894f1994723788c
(cherry picked from commit 3ff737d9b7c26962b7d12077530c48a87fe08fa2)