Fixing an issue if errors to the console are coming from non-EDT.
If error to be printed in the run console are coming from non-EDT, the
current window-showing code will assert. This fixes this issue by
guaranteeing that the window showing will only be invoked on the EDT
when it is not already showing.

Change-Id: I7bb122caea8523dc0e5f8fd196cdabaed9853717
(cherry picked from commit a6132fb5f1ec90b107d6c92d4536c819a43982ec)