Select devices in getState, not during execution.
This will allow device-specific builds in pre-run tasks, and cleans up
the code by breaking out some of the many responsibilities in
AndroidRunningState.

This change also makes launches easier and faster: After run/debug is
clicked, rather than waiting for the build to complete to pop a dialog
for the user to pick a device, we ask them immediately. The launch then
proceeds in the background without need for further interaction. If an
emulator launch is needed, it runs in parallel with the build, speeding
up the launch.

This change also moves device launch/selection closer to extensibility.

Change-Id: I1f2c0180198f96838e4639efda076f4347470669