Fix InferNullity to be able to handle a project with multiple modules.
Only the modules that need to use annotations should be modified
to include the annotation dependency. Also if AppCompat is already
an dependency we do not need to add the annotation dependency.
This is a fix for Issue: 185988

Change-Id: Ieae69e5cbf3c6b3321ee2d81b75bf6808d9073dd
(cherry picked from commit 8f855d030d007db558f8bf85dad4160f41697f81)