Show disabled themes in the theme selector
When a theme is not available at the current configuration,
still show it in the theme selector, but make it non-selectable.
Also disable the attributes table when switching to a configuration
where the current theme is not defined.

Bug: http://b.android.com/182423
Change-Id: If3f1071d52823c56b34fafba204547d5a414bbb8
(cherry picked from commit d0b77d64a9ff5c67d14703a3e5e150c868896024)