Created main frame for the new "Project Structure" dialog.
This CL starts the implementation of the new "Project Structure"
dialog. Its purpose is to drive the development of the new
build.gradle parser, since this dialog is the ultimate test
for the parser.

This implementation only contains the following:
- Borrows the L&F of the new "Project Structure" dialog in
  IDEA 15
- Displays the "SDK Paths" configuration page
- Includes an extension point for specifying extra
  "configurables" groups
- Adds an implementation of the extension points for Gradle-based
  modules (it only implements the infrastructure for implementing
  tabbed editors)
- Implements navigation history

This new PSD is available by passing the system property
"-Duse.new.project.structure=true" to Android Studio.

Change-Id: Ief746349a69603e2fbba32912b93d0c6fc29254f