Adding bitmap preview to hprof viewer.
This CL pulls the external contribution from aosp CL 169293. The CL
includes major refactoring to share code with the debugger popup
bitmap viewer. It also includes various changes/cleanups to fit the
coding standards and use more efficient methods.

A big thanks to the external contributor for getting this rolling.

Change-Id: I9e271523e66bb8b046554b188c048e865e52da64