Adding Generic Gradle Parser.
Presenlty this parser handles the ext, android android.defaultConfig
and android.productFlavors blocks, and it can be easily extended to
parse the other remaining blocks.

Adding ExtModel, AndroidModel, ProductFlavors and ProductFlavor to
handle the ext, android, productFlavors and defaultConfig blocks.

Adding generic LiteralElement, ListElement and MapElemet classes
to store the cotents of a GrLiteral, list of LiteralElemets and
a map from properties of type String to the values of type
LieralElement.

Change-Id: Id135c0876275e5565ce1c640219a95c55ff268ce