Use full configuration when resolving attributes/parent
This CL updates the resolution code to use a full configuration to use
the most specific configuration possible. This avoids in most cases
configurations matching multiple folders and returning the incorrect
result.

For example, for a -port configuration, it would match both default and
v21 since it doesn't define any version. Using the full configuration it
would match v21 since it's more specific.

Bug: http://b.android.com/186550

Change-Id: Ie0dc631fdbe0dc8eb1ff04bbcb159f39b314542c