RenderClassLoader: Be dumb in locating class file.
The class name passed to a ClassLoader must always be unabmiguous. An
inner class should be referenced using a '$' symbol (eg. foo.Bar$Baz).
There's no need to be smart and try to resolve foo.Bar.Baz to
foo.Bar$Baz. If the ClassLoader is getting ambiguous class names,
there's a bug else where that should be fixed. However, no such bug was
found in the testing done.

Change-Id: Ie2c7af5b5b8274bcf9339d0a51cfd67ed932597c