Fix: Cannot create a new module
An assertion in the template code was blocking new module creation.
The problem was that a module is found, but it has not been properly
setup yet i.e. AndroidFacet is null.
The fix is to use the default theme if this happens.

Change-Id: Ib5d8e8395559e00a6f010cb614020b67a645fd44
(cherry picked from commit 73bb54a28facd2f2c26230b95ef87309308babb1)