Implementation of the memory view.
When a memory range is selected in the atom tree, the memory view
will fetch and display the memory contents for that range.

Although the current version fetches and formats the entire
requested memory range at once, the MemoryModel class can be altered
to only fetch and format data on demand.

Change-Id: I7af598436a0ca1af64e02c76f00acf86fd5af3b3