Simplify checking for whether app can be debugged
This CL reorganizes (and simplifies) the code that checks for
whether an app is debuggable. Earlier, this used to be passed
in as one of the long list of parameters to AndroidRunningState.
Now we compute it directly in the AndroidRunningState when it
is needed. This CL also folds in the method prepareAndStartApp
into doPrepareAndStart.

Change-Id: Ia776827b70fb245f20611c46777dbb865732ac18