Adds a default Captures tool window order attribute as 2.
There was no default order attribute for Captures window, and that
makes the Captures window be above Project and Structure windows
in a new project. When a project component ToolWindowManager is
being initialized, the ComponentStoreImpl is looking for the default
config, see ComponentStoreImpl#getDefaultState. That method queries
resources to get a file that matches with the component name
"/idea/ToolWindowManager.xml" or "ToolWindowManager.xml". The new xml
is similar to WindowManager.xml.

Change-Id: I2e63599f328ac1d83631b93838b52d1e906b050d