185634: Fix styles parent rename
This fixes the rename case where a child style defines a parent by name
"MyParentName.MyChildName" (instead of using the "parent" attribute) and
we rename "MyParentName".
The CL will find all the instances where there are references to
"MyParentName" and replace them with the new name.
This also fixes the case where "MyParentName.MyChildName" was being
referenced from other attributes like the theme attribute in the
manifest.

Bug: http://b.android.com/185634
Change-Id: I067943af11b6dbd96fb2741ea7d923f025179e00