Increment format version and add code to support the old format.
This is a fix for issue: 187515
We changed the way templates should be setup in 1.4 in that recipe
files now must specify the implicit "root" folder. This CL ensures
that old templates can be loaded - which is required for 3rd party
templates.

Cherry pick of: Ieaeb6f305a90d88f3b732c947bf98dd5044e60ff

Change-Id: I8e0afe101402eb0ae9587c8704912f3f1b546428