Collect config errors rather than throwing.
When we threw an exception for any warning or fatal error we
encountered, we introduced a lot of cases where fatal errors were masked
by warnings. Making sure that all the fatal checks are performed before
all the warning checks is not really feasible, because the validation of
different pieces of state is delegated to many different methods.

Instead, we collect and return all issues, and surface the most severe
to the user.

Change-Id: I524b6d0e88a9ff04c37d41eadf256e43abe3cfbb