Use options object for running state settings.
Replace a couple AndroidRunningState constructor parameters, setters,
and run configuration reads with an immutable options object. Allow
AndroidRunConfigurationBase subclasses to adjust the options before
they are passed to AndroidRunningState.

We could add more to the options object, but none of the other
AndroidRunningState construction parameters have a reasonable
"default" value, so they would probably all have to be construction
parameters to the builder anyway.

Change-Id: Iffb688e2d536f0a867a91002de57e1fca2bc6f10