Now "Run Configurations" are created on project creation.
Run configurations were not created when an Android project
was created. The cause was a change introduced in IDEA 15
where any modifiable model is not committed until the
Gradle project is fully customized. Because of this,
AndroidFacet was not immediately added to a module in
AndroidFacetModuleCustomizer, resulting in
RunConfigurationModuleCustomizer not being able to find an
AndroidFacet in a given module and skipping the creation
of the "Android" run configuration.

Change-Id: Ief318fee6ef88ee54df544abce31b3e12e11a824