Remove alarm in device chooser, use callbacks
The DeviceChooser uses an alarm (500ms) to poll for devices and
updates the table as needed. Since the alarm goes all the time, the
actual table is only refreshed only when the list of device changes.
This is problematic if the list of devices doesn't change, but the
state of one particular device changes.

This CL removes the entire alarm based architecture, and instead uses
callbacks from AndroidDebugBridge. The number of updates is now
significantly less, and this allows us to refresh the table
everytime there is a real change. We now also listen to bridge
connection changes, which means that killing adb server while the
dialog is open results in the appropriate behavior.

Fixes https://code.google.com/p/android/issues/detail?id=188105

Change-Id: I6c3a9b91e6becbefac2cf2ab12da672b8222b4c2