Extract out some common logic for a device specific cache
The InstalledApkCache contains logic for a cache that maintains
a mapping of <device, pkg> to some specific state. This CL extracts
just that caching logic to a separate class in anticipation of its
use in a different place (we need to maintain a map of installed
dex files per device).

Change-Id: I1a3855c96cc7f3579f5ed4f2003fc562fadc7344