Refactor template code to pass a RenderingContext.
This is a preparation for adding a dryRun option.
We can add this option to RenderingContext without having to add
yet another parameter to a lot of methods.
Also move the file gatherings from recipe into this context. There
are now fewer places where we need to propagate open files and
target files.

Change-Id: I06b7705c9e42026de755ab72048213761e294100