Remove the setting of 'relativePackage' for new project.
This was added here: Iae29482e27870ce5449717f10ec1c31c20801674
However I the only effect I see is that the tools:context tag
doesn't have the package name for "New Project" but it is
available for "Add Activity". I removed it such that these are
the same. This eliminates warnings when adding a xml file twice
(as it can happen for new master detail activity).

Change-Id: I9629b802cb80f5ad4764a72017da4e1df6114d0b