Always do a full launch for debug builds
The launch infrastructure doesn't currently support incremental debug
action. A debug launch will always be a full install. This will be fixed
shortly, but until it is fixed, this CL will make it so that the
behavior is predictable: you can debug and then do multiple incremental
runs while debugging, but if you press debug again, it will be a full
launch.

Change-Id: I29bf71a9cc254360caa34d6297c8d1c658752184