Move Patch updates to separate classes
Rather than having a single program runner, we now have multiple runners
depending on the current state. The first time an app is deployed, the
run configuration returns an AndroidRunningState. If the run
configuration instead detects that a patch update can be performed, it
returns a PatchDeployState, which also results in the usage of the
PatchRunner instead of AndroidDebugRunner.

This cleanly separates patch update logic from the rest of the launch
config state: We detect really early whether this is a patch flow or
not, and everything else (the runner, and the run profile state) is now
different based on that choice.

Change-Id: I5186950a12e2eb67c89c674242cb3f57c1140605