Perform a full rebuild if manifest contents changes
If you edit the manifest file, we need to perform a
complete rebuild and reinstall. This CL keeps track
of the manifest timestamps when APK's are pushed to
a device, and then on incremental build time, checks
whether there is an edit in any of the project manifest
files and if so, triggers a full build.

Change-Id: Ife8cadf1d9f388e275e78fd9276a99d2731e8253