186778: Do not use module after being disposed
When a module is renamed, the old version is disposed. This causes a
crash in AppResourceRepository since it keeps a reference to the
module. When the project is rebuilt, the build notification is sent
and produces the crash.
This fixes the problem by attaching the message bus subscription
lifetime to the module lifetime.

Bug: http://b.android.com/186778
Change-Id: Ida9d21786faa6eea65b90b315099414fa36eee06