189658: Do not attempt to validate if device chooser is not enabled
The device chooser table runs its own validation routines and shows
errors if the selected devices aren't compatible. However, the device
chooser can be embedded in a different dialog, and in that context, can
sometimes be disabled. It should not run its validation routines in such
a case.

This partly fixes https://b.android.com/189658, but a complete fix that
validates the selected avd will have to wait for the device chooser
redesign.

Change-Id: I0a6f54e8df84e51b063ffa75792314242aa07631