Update RenderingContext to use custom FreemarkerConfiguration
Our codebase uses a custom FreemarkerConfiguration which, among other
differences from the default configuration, includes knowledge of
special objects which it can then convert to string, int, and bool
types.

After a refactoring, this dependency was accidentally lost, causing
our template system to throw a NotBooleanException when processing a
line like the following:

<#if some.bool.value>

which, due to the misconfiguration, caused the value stored at that
variable to be interprested as a string value and not a bool.

Change-Id: I1e0cfa8631df47e4c247b9e87622358c794301c0