Only add references when we know a file was generated.
If we decide not to copy a file we do not want to add it into the
target files collected.
By doing this we avoid reformatting files that we did not change.
This was a problem for the dummy data class in "Masterdetail"
since the reformatter is not idempotent.
Also: make all file references that are collected absolute.
This will make it easier to use afterwards.

Change-Id: I0938b1582a5f5b8dcba1ca7dc4b082751484e835